Backend

See merge request s07-webmobile1-sub2/S07P12D201!40

+++ backend/src/main/java/com/lemonmul/planetdhance/entity/Clear.java
@@ -1,5 +1,6 @@
 package com.lemonmul.planetdhance.entity;
 
+import com.lemonmul.planetdhance.entity.user.User;
 import lombok.*;
 
 import javax.persistence.*;
@@ -10,10 +11,9 @@ import javax.persistence.*;
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 public class Clear {
     @Column(columnDefinition = "INT UNSIGNED", name ="clear_id")
-//    @Column(name ="clear_id")
     @Id
     @GeneratedValue
-    private int id;
+    private Long id;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "music_id")
